modify. daisyUI,write

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -1,24 +1,30 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-<c:set var="pageTitle" value="Detali" />
+<c:set var="pageTitle" value="Detail" />
 <%@ include file="../common/head.jsp" %>
 	<section class="mt-8 text-xl">
 		<div class="container mx-auto px-3">
 			<div class="table-box-type-1">
-				<table>
+				<table class="table table-zebra">
+					<colgroup>
+						<col width="200" />
+					</colgroup>
 					<tbody>
-						<c:forEach var="article" items="${article }"  var="loginedMemberId" items="${loginedMemberId }">
 						<tr>
 							<th>번호</th>
-							<td>${article.id }</td>
+							<td><div class="badge">${article.id }</div></td>
 						</tr>
 						<tr>
-							<th>작성 날짜</th>	
+							<th>작성날짜</th>
 							<td>${article.regDate }</td>
 						</tr>
 						<tr>
-							<th>작성자
+							<th>수정날짜</th>
+							<td>${article.updateDate }</td>
+						</tr>
+						<tr>
+							<th>작성자</th>
 							<td>${article.writerName }</td>
 						</tr>
 						<tr>
@@ -26,21 +32,20 @@
 							<td>${article.title }</td>
 						</tr>
 						<tr>
-							<th>내용
+							<th>내용</th>
 							<td>${article.body }</td>
 						</tr>
-					</c:forEach>
 					</tbody>
 				</table>
 			</div>
-			<div class="btns">
-				<button class="btn-text-link" type="button" onclick="history.back();">뒤로가기</button>
-			    <c:if test="${article.id == loginedMemberId}">
-					<a class="btn-text-link" href="modify?id=${article.id }">수정</a>
-					<a class="btn-text-link" href="doDelete?id=${article.id }">삭제</a>
-   				 </c:if>
-			</div>
 			
+			<div class="btns mt-2">
+				<button class="btn-text-link btn btn-active btn-ghost" type="button" onclick="history.back();">뒤로가기</button>
+				<c:if test="${article.actorCanChangeData }">
+					<a class="btn-text-link btn btn-active btn-ghost" href="modify?id=${article.id }">수정</a>
+					<a class="btn-text-link btn btn-active btn-ghost" href="doDelete?id=${article.id }" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+				</c:if>
+			</div>
 		</div>
 	</section>
 <%@ include file="../common/foot.jsp" %>
